3.0 liaison updates

+++ src/util/ais/encode.cpp
@@ -164,7 +164,8 @@ void goby::util::ais::Encoder::encode_msg_18(const goby::util::ais::protobuf::Po
         {30, static_cast<std::uint32_t>(pos.mmsi())},        // mmsi
         {8},                                                 // regional reserved
         {10, ais_speed(pos.speed_over_ground_with_units())}, // sog in 1/10 knots
-        {1, pos.position_accuracy()},           // position accuracy (0 = GNSS fix), (1 = DGPS)
+        {1, static_cast<std::uint32_t>(
+                pos.position_accuracy())},      // position accuracy (0 = GNSS fix), (1 = DGPS)
         {28, ais_latlon(pos.lon_with_units())}, // 1/10000 minutes
         {27, ais_latlon(pos.lat_with_units())}, // same as lon
         {12, pos.has_course_over_ground() ? ais_angle(pos.course_over_ground_with_units(), 1)
@@ -215,7 +216,7 @@ void goby::util::ais::Encoder::encode_msg_24(const goby::util::ais::protobuf::Vo
             {2},                                          // repeat indicator
             {30, static_cast<std::uint32_t>(voy.mmsi())}, // mmsi
             {2, part_num},                                // part num
-            {8, voy.type()},                              // ship type
+            {8, static_cast<std::uint32_t>(voy.type())},  // ship type
             {18, 0, "XXX", true},                         // vendor ID
             {4, 0},                                       // unit model code
             {20, 0},                                      // serial number
